% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-documentation.R, R/rzig-wrappers.R
\name{poc_single_zig}
\alias{poc_single_zig}
\title{Compute one single-term Li-Pearl probability-of-causation interval.}
\usage{
poc_single_zig(kind, aji, oji, vi, vk, uj, up, opk, ojk, other_ok)
}
\arguments{
\item{kind}{Theorem number, 4 through 7.}

\item{aji}{Interventional probability for the target treatment and outcome.}

\item{oji}{Observed joint probability for the target treatment and outcome.}

\item{vi}{Observed marginal probability of the target outcome.}

\item{vk}{Observed marginal probability of the factual outcome.}

\item{uj}{Observed probability of the target treatment.}

\item{up}{Observed probability of the factual treatment.}

\item{opk}{Observed joint probability of factual treatment and outcome.}

\item{ojk}{Observed joint probability of target treatment and factual outcome.}

\item{other_ok}{Observed factual-outcome probabilities for other treatments.}
}
\value{
Lower and upper bounds.
}
\description{
Compute one single-term Li-Pearl probability-of-causation interval.
}
\examples{
# Theorem 4 for a binary synthetic table.
poc_single_zig(4L, .3, .15, .45, .45, .5, .5, .15, .15, .3)
}
\references{
Li, A. and Pearl, J. (2024). Probabilities of Causation with Nonbinary
Treatment and Effect. \emph{Proceedings of the AAAI Conference on Artificial
Intelligence}, 38(18), 20465--20472. \doi{10.1609/aaai.v38i18.30030}.

See \link{pnsbib-methods} for the method-to-function map, assumptions and
references. Graphics and reporting functions preserve supplied results;
they do not perform causal identification.
}
